+++ addOns/callhome/src/main/java/org/zaproxy/addon/callhome/ExtensionCallHome.java
@@ -290,6 +290,7 @@ public class ExtensionCallHome extends ExtensionAdaptor
                     || key.startsWith("stats.auth.")
                     || key.startsWith("stats.auto.")
                     || key.startsWith("stats.break.")
+                    || key.startsWith("stats.client.")
                     || key.startsWith("stats.code.")
                     || key.startsWith("stats.exim.")
                     || key.startsWith("stats.fuzz.")